requirements.txt
flask==2.3.3
openai==1.12.0
python-dotenv==1.0.0
moviepy==1.0.3
Pillow==10.0.0
tenacity==8.2.3
flask-cors==4.0.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
numpy==1.24.3
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9